modify size_t and int type

+++ include/common/TreeNode.hpp
@@ -45,8 +45,8 @@ Node *string2tree(std::string data) {
   Node *root = nullptr;
   std::queue<Node *> q;
   int count = 0;
-  for (int i = 1; i < data.size() - 1; ++i) {
-    int j = i;
+  for (size_t i = 1; i < data.size() - 1; ++i) {
+    size_t j = i;
     while (j < data.size() - 1) {
       if (data[j] == ',' or data[j] == ']') break;
       ++j;
@@ -129,7 +129,7 @@ void print_tree(Node *root) {
     }
     delete null_node;
     std::cout << v[0];
-    for (int i = 1; i < v.size(); ++i) {
+    for (size_t i = 1; i < v.size(); ++i) {
       if (v[i] == -1)
         std::cout << ",null";
       else
