+++ src/Apr/4.27/136.cpp
@@ -14,7 +14,7 @@ public:
     int singleNumber(vector<int> &nums)
     {
         int result = 0;
-        for (int i = 0; i < nums.size(); ++i)
+        for (size_t i = 0; i < nums.size(); ++i)
         {
             result ^= nums[i];
         }
