+++ src/May/5.12/205.cpp
@@ -14,9 +14,9 @@ class Solution {
     char map_t[128]{0};
     char map_s[128]{0};
     for (size_t i = 0; i < s.size(); ++i) {
-      if (not map_t[s[i]]) map_t[s[i]] = t[i];
-      if (not map_s[t[i]]) map_s[t[i]] = s[i];
-      if (map_t[s[i]] != t[i] or map_s[t[i]] != s[i]) {
+      if (not map_t[(uint8_t)s[i]]) map_t[(uint8_t)s[i]] = t[i];
+      if (not map_s[(uint8_t)t[i]]) map_s[(uint8_t)t[i]] = s[i];
+      if (map_t[(uint8_t)s[i]] != t[i] or map_s[(uint8_t)t[i]] != s[i]) {
         return false;
       }
     }
